+++ bcipy/helpers/list.py
